requirements.txt
streamlit==1.33.0
python-dotenv==1.0.0
requests==2.31.0
openai==1.55.3
folium==0.14.0
streamlit-folium==0.15.1